+++ .gitignore
@@ -10,6 +10,7 @@
 /build
 
 # misc
+.idea
 .DS_Store
 .env.local
 .env.development.local
