+++ src/actions/actions.js
@@ -1,0 +1,5 @@
+import * as characterActions from './characterActions/characterActions';
+
+export {
+  characterActions
+};
