+++ src/actions/characterActions/characterActions.spec.js
@@ -1,12 +1,13 @@
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
+
 import {
   setCharacters,
   upvoteCharacter,
   getCharacters,
   SET_CHARACTERS,
   UPVOTE_CHARACTER,
-} from '.';
+} from '../actions';
 
 describe('characterActions', () => {
   describe('setCharacters', () => {
