--- src/actions/constants.js
@@ -1,2 +1,0 @@
-export const SET_CHARACTERS = 'SET_CHARACTERS';
-export const UPVOTE_CHARACTER = 'UPVOTE_CHARACTER';
