--- src/actions/index.js
@@ -1,7 +1,0 @@
-export {
-  getCharacters,
-  setCharacters,
-  upvoteCharacter,
-} from './characterActions';
-
-export { SET_CHARACTERS, UPVOTE_CHARACTER } from './constants';
