+++ src/components/App/App.spec.jsx
