+++ src/components/CharacterList/CharacterList.spec.jsx
