+++ src/components/components.js
@@ -1,0 +1,2 @@
+export { default as App } from './App/App.jsx';
+export { default as CharacterList } from './CharacterList/CharacterList.jsx';
