+++ src/containers/containers.js
@@ -1,0 +1,1 @@
+export { default as AppContainer } from './AppContainer/AppContainer';
