+++ src/reducers/charactersReducer/charactersReducer.js
@@ -1,4 +1,6 @@
-import { SET_CHARACTERS, UPVOTE_CHARACTER } from '../actions';
+import { characterActions } from '../../actions/actions';
+
+const { SET_CHARACTERS, UPVOTE_CHARACTER } = characterActions;
 
 const upvoteCharacter = (state, character) => {
   const newState = [...state];
@@ -6,7 +8,8 @@ const upvoteCharacter = (state, character) => {
 
   if (newState[index].voteCount) {
     newState[index].voteCount += 1;
-  } else {
+  }
+  else {
     newState[index].voteCount = 1;
   }
   return newState.sort((a, b) => a.voteCount < b.voteCount);
