+++ src/reducers/charactersReducer/charactersReducer.spec.js
@@ -1,0 +1,1 @@
+// todo
