+++ src/reducers/index.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
-import characters from './charactersReducer';
+import characters from './charactersReducer/charactersReducer';
 
 export default combineReducers({
-  characters,
+  characters
 });
