+++ src/store/store.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-underscore-dangle */
 import { createStore, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
-import rootReducer from './reducers';
+import rootReducer from '../reducers/index';
 
 const initialState = {};
 const enhancers = [];
